refactor: move ChangeProof from merkle/changes to proofs/change
